[action] [PR:20635] Fix UnboundLocalError in lldp test (#782)
Initialize lldpctl_interface in ansible/tests/lldp/test_lldpsyncd.py in order to prevent potential UnboundLocalError.

<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit easier:
-->
### Description of PR
<!--
- Please include a summary of the change and which issue is fixed.
- Please also include relevant motivation and context. Where should reviewer start? background context?
- List any dependencies that are required for this change.
-->

Summary:
Fixes # (issue)

Fix potential UnboundLocalError in ansible/tests/lldp/test_lldp_syncd.py
by initializing lldpctl_interface to None. The assertion this variable
is passed to performs an existence check, and so will fail the testlet
appropriately if lldpctl_interface is not set.

### Type of change

- [x] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
 - [ ] Skipped for non-supported platforms
- [ ] Test case improvement

### Back port request
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x] 202505

### Approach
#### What is the motivation for this PR?
Enhance the debug-ability of ansible/tests/lldp/test_lldpsyncd.py such that in the even that no lldp interface is identified in the test, the error is propagated up to and reported by the verification method and does not simply throw UnboundLocalError.

#### How did you do it?
Fix is to initialize the variable lldpctl_interface to None outside of conditional logic so that it has a value when passed to the test verification method.

#### How did you verify/test it?
Verified that the same failing test case no longer throws UnboundLocalError and fails in a meaningful way for further debug.