Increase the pre_reboot_timeout for fwutil ONIE/BIOS test for SN4280 (#801)
Description of PR
Increase the pre_reboot_timeout for fwutil ONIE/BIOS test for SN4280


What is the motivation for this PR?
The switch shutdown time during the auto reboot fwutil tests(ONIE/BIOS)
is longer due to
PR:https://github.com/sonic-net/sonic-buildimage/pull/24165. We need
increase the timeout.
This sonic-buildimage PR#24165 will be removed in 202511 and master
branch later, so the test change is only for 202506.

How did you do it?
Increase the timeout by additional 360s which is the same timeout used
in DPU reboot tests for a DPU reboot.

How did you verify/test it?
Run the ONIE/BIOS test cases with Nvidia SN4280, all passed.