fib/test_fib.py: Use IPv6 addresses in outer header for IPinIPHashTest (#917)
### Description of PR
When running on isolated-v6 testbeds, use IPv6 addressing in the
outermost L3 header of IPinIP packets, as IPv4 addresses are not
configured and are unresolvable on these testbeds.

### Type of change

- [x]  Bug fix
- [ ]  Testbed and Framework(new/improvement)
- [ ]  New Test case
- [ ]  Skipped for non-supported platforms
- [ ]  Test case improvement

### Approach
What is the motivation for this PR?
This removes an issue caused by a mismatch between test expectations and
the testbed configuration, namely the assumption that IPv4 addresses can
be used on testbeds that are isolated-v6. This came up during imagequal
testing of a different branch and will need to be addressed there as
well.

### How did you do it?
Evaluate whether or not the testbed is isolated-v6 and safe this in a
class attribute. If this attribute is true, pass IPv6 addresses to
packet-creation and verification methods so that the L3 addresses in
outer headers of IPinIP packets are IPv6. IPv6-specific scapy methods
are utilized where applicable.

### How did you verify/test it?
Verified that the test case now uses IPv6 addressing in outer IPinIP
headers, and that it no longer fails due to nexthop resolution issues.